Add code to allow adding HTML block level DIV tags and inline SPAN tags to be used to set 	text language  (HTML lang attribute) 	text direction (HTML dir  attribute) 	CSS class names in the HTML class attribute) and with special case the HTML style attribute.
The attribute values to be used are passed into openSpan() and openDiv()
using an options hash. See the code for the details.

There are matching closeSpan() and closeDiv() functions to close these tags.

All these functions currently depend on the author adding TeX or preTeXt
strings which should be added to the output, if needed. These default to being
empty strings at present.

The openSpan / closeSpan is more powerful than the $BLTR and $ELTR formatting
variables which only create a span to set the text direction to LTR inside a
problem rendered in RTL mode (based on course or question settings).